refactor: add modules folder and add auth module
