+++ src/modules/albums/interfaces/album.interface.ts
