+++ src/modules/artists/interfaces/artist.interface.ts
