+++ src/modules/favs/interfaces/fav-body.interface.ts
