+++ src/modules/favs/interfaces/fav.interface.ts
